backend/pyproject.toml
[project]
name = "ai-co-founder-for-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "feedparser>=6.0.12",
    "fpdf2>=2.8.4",
    "matplotlib>=3.10.6",
    "openai-agents>=0.2.11",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "playwright>=1.55.0",
    "plotly>=6.3.0",
    "praw>=7.8.1",
    "python-docx>=1.2.0",
    "reportlab>=4.4.3",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
